refactor: extract session + session_manager to internal/muxcore/session/ (T015)
